extracted utils and switch converter from server.js
server.js is way to big. Makes it hard to handle and error prone. So I started careful refactoring.

(By doing so I detected an unused paramter in _processCommon :smile: ).
